changes related to view employee screen

+++ src/app/employee/viewemployees/viewemployees.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { ChangeDetectorRef, Component } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import {
   BankDetailViewDto, EmployeAdressViewDto, EmployeeBasicDetailDto, EmployeeBasicDetailViewDto, employeeEducDtlsViewDto,
@@ -8,7 +8,7 @@ import { EmployeeService } from 'src/app/_services/employee.service';
 import { AssetAllotmentViewDto } from 'src/app/_models/admin/assetsallotment';
 import { AdminService } from 'src/app/_services/admin.service';
 import { MEDIUM_DATE } from 'src/app/_helpers/date.formate.pipe';
-import { Actions, DialogRequest, ViewEmployeeScreen } from 'src/app/_models/common';
+import { Actions, ConfirmationRequest, DialogRequest, ViewEmployeeScreen } from 'src/app/_models/common';
 import { AddassetallotmentDialogComponent } from 'src/app/_dialogs/addassetallotment.dialog/addassetallotment.dialog.component';
 import { UnassignassetDialogComponent } from 'src/app/_dialogs/unassignasset.dialog/unassignasset.dialog.component';
 import { DialogService, DynamicDialogRef } from 'primeng/dynamicdialog';
@@ -23,6 +23,7 @@ import { EducationdetailsDialogComponent } from 'src/app/_dialogs/educationdetai
 import { ExperiencedetailsDialogComponent } from 'src/app/_dialogs/experiencedetails.dialog/experiencedetails.dialog.component';
 import { ALERT_CODES, AlertmessageService } from 'src/app/_alerts/alertmessage.service';
 import { JwtService } from 'src/app/_services/jwt.service';
+import { ConfirmationDialogService } from 'src/app/_alerts/confirmationdialog.service';
 
 @Component({
   selector: 'app-viewemployees',
@@ -75,7 +76,8 @@ export class ViewemployeesComponent {
   dialog: boolean = false;
   empbasicDetails = new EmployeeBasicDetailDto();
   selectedOption: boolean;
-
+  confirmationRequest: ConfirmationRequest = new ConfirmationRequest();
+  
   constructor(
     private jwtService: JwtService,
     private alertMessage: AlertmessageService,
@@ -83,7 +85,8 @@ export class ViewemployeesComponent {
     private activatedRoute: ActivatedRoute,
     private adminService: AdminService,
     public ref: DynamicDialogRef,
-    private dialogService: DialogService) {
+    private dialogService: DialogService,
+    private confirmationDialogService: ConfirmationDialogService,) {
     this.employeeId = this.activatedRoute.snapshot.queryParams['employeeId'];
   }
 
@@ -153,21 +156,23 @@ export class ViewemployeesComponent {
   initUploadedDocuments() {
     this.employeeService.GetUploadedDocuments(this.employeeId).subscribe((resp) => {
       this.UploadedDocuments = resp as unknown as any[];
-      console.log(this.UploadedDocuments);
-
     });
   }
 
   removeItem(uploadedDocumentId: any) {
-    debugger
-    this.employeeService.DeleteDocument(uploadedDocumentId).subscribe(resp => {
-      console.log(resp);
-      
-        if (resp)
-            this.alertMessage.displayAlertMessage(ALERT_CODES["EAD006"]);
-        else
+    this.confirmationDialogService.comfirmationDialog(this.confirmationRequest).subscribe(userChoice => {
+      if (userChoice) {
+        this.employeeService.DeleteDocument(uploadedDocumentId).subscribe(resp => {
+          if (resp){
+              this.alertMessage.displayAlertMessage(ALERT_CODES["EAD006"]);
+              this.UploadedDocuments = this.UploadedDocuments.filter(doc => doc.uploadedDocumentId !== uploadedDocumentId);
+          }
+          else{
             return this.alertMessage.displayErrorMessage(ALERT_CODES["EAD007"]);
-    })
+          }         
+      })
+      }
+    });
 }
   downloadItem(uploadedDoucment) {
